refactor: 레이아웃 상태 전역상태로 관리
